requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==3.0.0
autopage==0.5.1
backcall==0.2.0
backoff==2.2.1
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
bleach==5.0.1
blis==0.7.9
boto3==1.26.17
botocore==1.29.17
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cliff==3.10.1
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
confection==0.0.3
configparser==5.3.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.3.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gdown==4.5.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
hmmlearn==0.2.8
huggingface-hub==0.10.1
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
julius==0.2.7
jupyter==1.0.0
jupyter-client==7.4.5
jupyter-console==6.4.4
jupyter-core==4.11.2
jupyter-server==1.23.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.2
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.7
notebook==6.5.2
notebook-shim==0.2.2
num2words==0.5.12
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.2.3
openai==0.25.0
openpyxl==3.0.10
optuna==3.0.3
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.62
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pbr==5.11.0
pickleshare==0.7.5
Pillow==9.3.0
pip-autoremove==0.10.0
pkgutil-resolve-name==1.3.10
pooch==1.6.0
preshed==3.0.8
prettytable==3.5.0
primePy==1.3
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
punctuator==0.9.6
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
PyMuPDF==1.21.1
pyparsing==3.0.9
pypdf2==2.12.1
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.19.2
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python3-pip-autoremove==0.9.1
pytorch-lightning==1.6.5
pytorch-metric-learning==1.6.3
pytube==12.1.2
pytube3==9.6.4
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shellingham==1.5.0
simplejson==3.18.0
singledispatchmethod==1.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sox==1.4.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
speechbrain==0.5.13
SQLAlchemy==1.4.44
srsly==2.4.5
stevedore==3.5.2
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
Theano==1.0.5
thinc==8.1.5
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.13.0+cu117
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.0
torchmetrics==0.10.3
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@9f70a352f9f8630ab3aa0d06af5cb9532bd8c21d
widgetsnbextension==4.0.3
yarl==1.8.1
zipp==3.10.0

text2emotion~=0.0.5
whisper~=1.0
LeXmo~=0.1.4
fastapi~=0.92.0